+++ packages/client-common/src/connection.ts
@@ -4,7 +4,7 @@ import type {
   WithHttpStatusCode,
   WithResponseHeaders,
 } from './clickhouse_types'
-import type { LogWriter } from './logger'
+import type { ClickHouseLogLevel, LogWriter } from './logger'
 import type { ClickHouseSettings } from './settings'
 
 export type ConnectionAuth =
@@ -19,6 +19,8 @@ export interface ConnectionParams {
   database: string
   clickhouse_settings: ClickHouseSettings
   log_writer: LogWriter
+  log_level: ClickHouseLogLevel
+  unsafeLogUnredactedQueries: boolean
   keep_alive: { enabled: boolean }
   application_id?: string
   http_headers?: Record<string, string>