+++ packages/client-node/__tests__/integration/node_exec.test.ts
@@ -1,4 +1,9 @@
-import type { ClickHouseClient } from '@clickhouse/client-common'
+import {
+  DefaultLogger,
+  LogWriter,
+  type ClickHouseClient,
+  ClickHouseLogLevel,
+} from '@clickhouse/client-common'
 import { describe, it, beforeEach, afterEach, expect } from 'vitest'
 import { createSimpleTable } from '@test/fixtures/simple_table'
 import { createTestClient } from '@test/utils/client'
@@ -10,8 +15,10 @@ import { getAsText } from '../../src/utils'
 
 describe('[Node.js] exec', () => {
   let client: ClickHouseClient<Stream.Readable>
+  let log_writer: LogWriter
   beforeEach(() => {
     client = createTestClient()
+    log_writer = new LogWriter(new DefaultLogger(), 'Connection')
   })
   afterEach(async () => {
     await client.close()
@@ -87,7 +94,15 @@ describe('[Node.js] exec', () => {
         values: stream,
       })
       // the result stream contains nothing useful for an insert and should be immediately drained to release the socket
-      await drainStream(execResult.stream)
+      await drainStream(
+        {
+          op: 'Insert',
+          query_id: execResult.query_id,
+          log_writer,
+          log_level: ClickHouseLogLevel.OFF,
+        },
+        execResult.stream,
+      )
       await checkInsertedValues([
         {
           id: '42',
@@ -112,7 +127,15 @@ describe('[Node.js] exec', () => {
       stream.push(null)
       // the result stream contains nothing useful for an insert and should be immediately drained to release the socket
       const execResult = await execPromise
-      await drainStream(execResult.stream)
+      await drainStream(
+        {
+          op: 'Insert',
+          query_id: execResult.query_id,
+          log_writer,
+          log_level: ClickHouseLogLevel.OFF,
+        },
+        execResult.stream,
+      )
       await checkInsertedValues([])
     })
 
@@ -131,7 +154,15 @@ describe('[Node.js] exec', () => {
         values: stream,
       })
       // the result stream contains nothing useful for an insert and should be immediately drained to release the socket
-      await drainStream(execResult.stream)
+      await drainStream(
+        {
+          op: 'Insert',
+          query_id: execResult.query_id,
+          log_writer,
+          log_level: ClickHouseLogLevel.OFF,
+        },
+        execResult.stream,
+      )
       await checkInsertedValues([
         {
           id: '42',
@@ -155,7 +186,15 @@ describe('[Node.js] exec', () => {
         values: stream,
       })
       // the result stream contains nothing useful for an insert and should be immediately drained to release the socket
-      await drainStream(execResult.stream)
+      await drainStream(
+        {
+          op: 'Insert',
+          query_id: execResult.query_id,
+          log_writer,
+          log_level: ClickHouseLogLevel.OFF,
+        },
+        execResult.stream,
+      )
       await checkInsertedValues([])
     })