+++ packages/client-node/__tests__/unit/node_client.test.ts
@@ -4,7 +4,11 @@ import type {
   BaseClickHouseClientConfigOptions,
   ConnectionParams,
 } from '@clickhouse/client-common'
-import { DefaultLogger, LogWriter } from '@clickhouse/client-common'
+import {
+  DefaultLogger,
+  LogWriter,
+  ClickHouseLogLevel,
+} from '@clickhouse/client-common'
 import { createClient } from '../../src'
 import {
   type CreateConnectionParams,
@@ -47,6 +51,8 @@ describe('[Node.js] createClient', () => {
       database: 'analytics',
       clickhouse_settings: {},
       log_writer: new LogWriter(new DefaultLogger(), 'Connection'),
+      log_level: ClickHouseLogLevel.OFF,
+      unsafeLogUnredactedQueries: false,
       keep_alive: { enabled: true },
       http_headers: {
         'X-ClickHouse-Auth': 'secret_token',